Improving puzzle definition and discovery (#10)
This removes the activation of each solver to find out which day it is a
solver for.

It also removes the settings IsActive and UseTestInput that was defined
in code for each solver. UseTestInput is now defined in console when
user runs solution. This makes it quicker to jump between using test
data and real data.